Added talk command's functions (basic), Changed nonce to long, added port number for users

+++ src/com/utd/ns/sim/server/Functions.java
@@ -15,14 +15,14 @@ import java.util.regex.Pattern;
  */
 public class Functions {
 
-    /**
-     * Decrements integer nonce by 1
+/**
+     * Decrements long nonce by 1
      *
      * @param nonce nonce of type int
      * @return decremented nonce as String
      */
-    public static String nonceFail(int nonce) {
-        return (Integer.toString(nonce - 1));
+    public static String nonceFail(long nonce) {
+        return (Long.toString(nonce - 1));
     }
 
     /**
@@ -32,17 +32,17 @@ public class Functions {
      * @return decremented nonce as String
      */
     public static String nonceFail(String nonce) {
-        return (nonceFail(Integer.parseInt(nonce)));
+        return (nonceFail(Long.parseLong(nonce)));
     }
 
     /**
-     * Increments integer nonce by 1
+     * Increments long nonce by 1
      *
      * @param nonce nonce of type int
      * @return incremented nonce as String
      */
-    public static String nonceSuccess(int nonce) {
-        return (Integer.toString(nonce + 1));
+    public static String nonceSuccess(long nonce) {
+        return (Long.toString(nonce + 1));
     }
 
     /**
@@ -52,7 +52,7 @@ public class Functions {
      * @return incremented nonce as String
      */
     public static String nonceSuccess(String nonce) {
-        return (nonceSuccess(Integer.parseInt(nonce)));
+        return (nonceSuccess(Long.parseLong(nonce)));
     }
 
     /**
@@ -62,8 +62,8 @@ public class Functions {
      * @param toWhat nonce sent
      * @return true or false
      */
-    public static boolean checkNonce(String what, int toWhat) {
-        return checkNonce(Integer.parseInt(what), toWhat);
+    public static boolean checkNonce(String what, long toWhat) {
+        return checkNonce(Long.parseLong(what), toWhat);
     }
 
     /**
@@ -73,7 +73,7 @@ public class Functions {
      * @param toWhat nonce sent
      * @return true or false
      */
-    public static boolean checkNonce(int what, int toWhat) {
+    public static boolean checkNonce(long what, long toWhat) {
         if (what == toWhat) {
             return true;
         } else {
